Added client and passed players to network by objects
